Began work on getting slurm working
