+++ .gitignore
@@ -5,3 +5,4 @@
 /server/data/
 /server/PeakLearnerSlurm.cfg
 /jbrowse/data/
+/runlog.txt
