+++ .travis.yml
@@ -1,12 +1,17 @@
 language: python
 
+sudo: required
+
+apt_packages:
+  - libdb5.3-dev
+  - slurm-wlm
+
 python:
   - 3.6
   - 3.7
   - 3.8
 
 before_install:
-  - sudo apt install libdb5.3-dev
   - python --version
   - pip install -U pip
   - pip install -U pytest
@@ -19,6 +24,11 @@ before_install:
   - pip install codecov
 
 script:
+  - sudo cp slurm.conf /etc/slurm-llnl
+  - sudo service munge start
+  - sudo /etc/init.d/slurmctld restart
+  - sudo /etc/init.d/slurmd restart
+  - sinfo
   - pytest tests/
 
 after_success:
