+++ slurm.conf
@@ -46,5 +46,5 @@ JobAcctGatherType=jobacct_gather/none
 #
 #
 # COMPUTE NODES
-NodeName=localhost CPUs=2 State=UNKNOWN
+NodeName=localhost Procs=16 Boards=1 SocketsPerBoard=1 CoresPerSocket=8 ThreadsPerCore=2 State=UNKNOWN
 PartitionName=debug Nodes=localhost Default=YES MaxTime=INFINITE State=UP
